modified structure according to mvvm
