+++ app/src/main/java/com/example/ticktickclone/models/CompletionStatus.kt
@@ -1,4 +1,4 @@
-package com.example.ticktickclone
+package com.example.ticktickclone.models
 
 enum class CompletionStatus {
     DONE, WONT_DO, NOT_MARKED
