+++ app/src/main/java/com/example/ticktickclone/models/TaskList.kt
@@ -1,0 +1,6 @@
+package com.example.ticktickclone.models
+
+data class TaskList(
+    var name: String,
+    val tasks: MutableList<Task> = ArrayList()
+)
