+++ app/src/main/java/com/example/ticktickclone/viewmodels/TaskListViewModel.kt
@@ -1,0 +1,11 @@
+package com.example.ticktickclone.viewmodels
+
+import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModel
+
+class TaskListViewModel(
+    private val savedStateHandle: SavedStateHandle
+) : ViewModel() {
+
+    //private val
+}
