+++ app/src/main/java/com/example/ticktickclone/views/CalendarFragment.kt
@@ -1,21 +1,21 @@
-package com.example.ticktickclone
+package com.example.ticktickclone.views
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.ticktickclone.databinding.StopwatchFragmentBinding
+import com.example.ticktickclone.databinding.CalendarFragmentBinding
 
-class StopwatchFragment: Fragment() {
+class CalendarFragment : Fragment() {
+    private lateinit var binding: CalendarFragmentBinding
 
-    private lateinit var binding: StopwatchFragmentBinding
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = StopwatchFragmentBinding.inflate(layoutInflater, container, false)
+        binding = CalendarFragmentBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
