+++ app/src/main/java/com/example/ticktickclone/views/StopwatchFragment.kt
@@ -1,21 +1,21 @@
-package com.example.ticktickclone
+package com.example.ticktickclone.views
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.ticktickclone.databinding.CalendarFragmentBinding
+import com.example.ticktickclone.databinding.StopwatchFragmentBinding
 
-class CalendarFragment: Fragment() {
-    private lateinit var binding: CalendarFragmentBinding
+class StopwatchFragment : Fragment() {
 
+    private lateinit var binding: StopwatchFragmentBinding
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = CalendarFragmentBinding.inflate(layoutInflater, container, false)
+        binding = StopwatchFragmentBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
